Record all channels of the matrix in both wav and raw mode
This is a commit with all the modifications needed in the c++ code from
tfg/ to be able to record all eight channels present in the matrix.

A function to record raw audio is added, due to errors in wav header function.
We need to capture the audio in a raw format and convert it to wav using
sox, because the matlab wav parser doesn't like our genrated files.
The command, assuming a 16000 Hz sampling frequency, should be
`sox -r 16000 -c 1 -e signed -c 1 -e signed -b 16 mic_16000_s16le_channel_0.raw channel_0.wav`

This commit also introduces a pair of parameters to choose the filename
and folder where the audio is recorded. To avoid accidental file
overwrites it always prepend a timestamp to all file names.

Rename num_channels -> num_channels_wav in write_wav_header to make it
clear that the parameter refer to the number of channels in the wav
file, and it has no relationship with the number of microphones in the
matrix.

This commit also charges the sleep time of the processing thread from 1
ms to a minimum determined by the block size of the audio samples, which
depends on the sampling frequency (at 16KHz is 32 ms).

The explicit file closing code is deleted and, instead now it relays on
the class destructor, which automatically closes the file handle.

Also delete dead code, in particular `thread_manager.{cpp,hpp}`.