Add a function to send data to a mqtt broker.
The sync interface requires that we have a connect and disconnect
functions in addition to a send one, as the expected workflow is:

```cpp
connect_sync_mqtt_client();
send_mqtt_sync();
send_mqtt_sync();
... and at the end ...
disconnect_sync_mqtt_client();
```
This way we avoid making a new connection to the server to send a new
data block.

We also provide the option to connect with specific settings, and if
none are given we use the default ones.

Note that to use you need to construct a client in your main function,
as we can't construct and copy one with return, as the mqtt::client
class it's non-copyable.