Add a running bool to the async queue.
Using wait_pop() is desirable, as it can suspend the reading thread when
the producer is behind the consumer, but with the current code we can
have a deadlock if we change running to false while the thread is
waiting.

This is because no more data is produced, so the condition variable is
not notified, and even if we have a sudden wakeup the condition
!queue.empty() is not going to trigger, so the thread is going to go
back to sleep.

There are two possible solutions:
- Don't use wait_pop(), and do busy looping with pop().
- Introduce a function to notify to the queue that we are not going to
  do more work.

We adopt the second solution, but now we have to notify if we have more
data or not so use a return value of 1, to indicate that we are running
and we have more data, -1 to signal that we are not running, but we are
returning one of the leftovers elements in the queue and 0 to tell the
caller that we are no longer running and we have no more data.