Make uart.cpp compile with new wiringPi versions.
This commit deletes the use of an active waiting for an interrupt (which
has been deprecated for a while) by using a condition variable to wake
up the tread when an ISR fires, as it's done in other parts of the
codebase.

As a pleasant side effect it also eliminates the use of
`std::system("gpio...)`, which doesn't work on modern raspbian systems.