Simplify async audio capture and recording.
This simplifies the audio capture by eliminating the concept of
duration, as now the function only knows about a atomic_bool&, used to
determine if it continues to run.

In the wav writing function the duration is also eliminated by keeping
track of how many samples we have recorded and rewriting the wav header
on each write.

In the previous implementation we could theoretically lose data because
running turned false before the queue was empty. To avoid it now there
is a new boolean to indicate if we want to drain the queue and record
any leftover elements from when the running atomic_bool& was true.

Lastly, the old guessing game of filenames to avoid overwriting files
is removed, aligning the function with the sync version. However there
is an important difference, as this async version overwrites any file
with the same name as the given one.