Save previous position of outstream when writing wav headers.
When updating a wav_file to add more data the two length files in the
header must be updated. For simplicity we can call write_wav_header with
the new size.

The problem is that the function does a seek to the beginning of the
file, and if we try to write again without this change we would
overwrite the first bytes of the wav data section.

Instead we save the current position of the output stream and restore it
when we finish writing the header, only if the previous position is not
0, as that would indicate that we have not written before.

The tellp function can also return -1 to indicate that the stream is not
valid, so we don't move the write cursor in that case