Use a per channel length count in record_all_channel_wav
Previously we were using a global counter for all channels, that would
reset in each operation of the while(running) loop.  That is obviously
incorrect because we were always putting a length of 2048 in the header.

This commit changes that to a using an array, instantiated outside of
the loop, to avoid those problems.